Support anonymous config

+++ src/main/perl/Configuration.pm
@@ -51,27 +51,39 @@ Module provides the Configuration class, to manipulate confgurations.
 my $ec = LC::Exception::Context->new->will_store_errors;
 
 my $PROFILE_DIR_N = "profile.";
-my $ACTIVE_FN     = "ccm-active-profile.";
 
-# in this hash we keep track number of open configuration with given
-# cid. it is used for creation and removal of pid_files
+=item new
 
-#
-# Create Configuration object. It takes three arguments:
-#   $cache_manager - CacheManager object
-#   $cid - configuration id
-#   $locked - true or false lock flag
-#
-# ccm-active-profile-$cid.$pid is created in profile.$cid directory (where pid is process
-# id).
-#
-# If configuration with specified cid does not exists exception is
-# thrown.
-#
+Create Configuration object. It takes three arguments:
+    C<cache_manager>: the CacheManager object
+    C<cid>: the configuration id
+    C<locked>: boolean lock flag
+    C<anonymous>: boolean anonymous flag
+
+If a configuration with specified CID does not exists, an exception is
+thrown.
+
+When the C<locked> flag is set (or when the C<lock> method is called at any point), 
+the Configuration instance is bound to the the specific CID, even if this is not 
+the latest one (or when a new one is fetched during the lifetime of the process).
+An unlocked Configuration instance will always use the latest CID.
+
+Unless the anonymous flag is set to true, each process that creates a 
+Configuration instance, creates a file named C<ccm-active-profile.$cid.$pid> 
+(with C<$cid> the CID and C<$pid> the process ID) under the C<profile.$cid> 
+directory in the C<CacheManager> cache path. The presence of this file protects 
+the process from getting this particular CID removed by the C<ccm-purge> command 
+(e.g. by the daily purge cron job).
+
+Processes that have no permission to create this file (or don't care about long 
+runtimes), can set the C<anonymous> flag and use the configuartion 
+(at their own risk).
+
+=cut
 
 sub new
 {    #T
-    my ($class, $cache_manager, $cid, $locked) = @_;
+    my ($class, $cache_manager, $cid, $locked, $anonymous) = @_;
     my $cache_path = $cache_manager->getCachePath();
     unless ($cache_path =~ m{^([-./\w]+)}) {
         throw_error("Cache path '$cache_path' is not an absolute path");
@@ -91,12 +103,14 @@ sub new
         "cache_path"    => $cache_path,
         "cfg_path"      => $cfg_path,
         "cid_to_number" => undef,
+        "anonymous"     => $anonymous,
     };
     unless (-d $cfg_path) {
         throw_error("configuration directory ($cfg_path) does not exist");
         return ();
     }
     bless($self, $class);
+
     unless ($self->_create_pid_file($self)) {
         $ec->rethrow_error();
         return ();
@@ -128,8 +142,9 @@ sub getCacheManager ()
 sub _pid_filename
 {
     my ($self, $cid) = @_;
+    
     $cid = $self->{"cid"} if (!defined($cid));
-    return $self->{"cfg_path"} . "/${ACTIVE_FN}${cid}-" . getpid();
+    return $self->{"cfg_path"} . "/ccm-active-profile.${cid}-" . getpid();
 }
 
 #
@@ -142,6 +157,9 @@ sub _create_pid_file
     my ($self) = @_;
     unless ($self->{"cid_to_number"}{$self->{"cid"}}) {
         $self->{"cid_to_number"}{$self->{"cid"}} += 1;
+
+        return SUCCESS if $self->{anonymous};
+
         my $pid_file = $self->_pid_filename();
         unless (_touch_file($pid_file)) {
             throw_error("_touch_file($pid_file)", $ec->error);
@@ -166,8 +184,10 @@ sub _remove_pid_file ()
     }
     $self->{"cid_to_number"}{$cid} -= 1;
     if ($self->{"cid_to_number"}{$cid} == 0) {
-        my $pid_file = $self->_pid_filename($cid);
 
+        return SUCCESS if $self->{anonymous};
+
+        my $pid_file = $self->_pid_filename($cid);
         if ((-f $pid_file) && !unlink($pid_file)) {
             throw_error("unlink($pid_file)", $!);
             return ();
